requirement.txt
asgiref==3.2.7
Django==2.2.11
djangorestframework==3.11.0
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.1
